# swagger/models/base_model.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// BaseModel base model
//
// swagger:model BaseModel
type BaseModel struct {

	// ok
	// Required: true
	Ok *bool `json:"ok"`

	// request id
	// Required: true
	// Format: uuid
	RequestID *strfmt.UUID `json:"request_id"`
}

// Validate validates this base model
func (m *BaseModel) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateOk(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateRequestID(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *BaseModel) validateOk(formats strfmt.Registry) error {

	if err := validate.Required("ok", "body", m.Ok); err != nil {
		return err
	}

	return nil
}

func (m *BaseModel) validateRequestID(formats strfmt.Registry) error {

	if err := validate.Required("request_id", "body", m.RequestID); err != nil {
		return err
	}

	if err := validate.FormatOf("request_id", "body", "uuid", m.RequestID.String(), formats); err != nil {
		return err
	}

	return nil
}

// ContextValidate validates this base model based on context it is used
func (m *BaseModel) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *BaseModel) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *BaseModel) UnmarshalBinary(b []byte) error {
	var res BaseModel
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}